Delete existing events before import
Delete any existing events in a calendar before importing it.
At the end, we should only have the events that still exist in Google
Calendar. They could have been moved or deleted, so don't keep them
around in Nextcloud. This is especially bad when they got moved like
half an hour and a bunch of events pile up on the same time slot.

This is not the most efficient way to do it, but it's the easiest.

Signed-off-by: Marcel Robitaille <mail@marcelrobitaille.me>