Properly track what events need to be deleted
Instead of deleting everything from the calendar being imported just to
add it all again, which does a LOT of slow DB writes, read all the
existing events from the database into a set. When considering one from
Google, check if it's in the set. If it is, remove it from the set and
skip importing it. Otherwise, add it. At the end, everything left in the
set was deleted from Google and must be deleted here.

This brings one of my >1000 event calendars from a 162 second import to
a 4 second import.

Signed-off-by: Marcel Robitaille <mail@marcelrobitaille.me>